moved MockWKScriptMessage into its own file
