+++ Tests/KlaviyoUITests/IAFWebViewModelPreloadingTests.swift
@@ -10,25 +10,6 @@ import KlaviyoCore
 import WebKit
 import XCTest
 
-class MockWKScriptMessage: WKScriptMessage {
-    private let mockName: String
-    private let mockBody: Any
-
-    init(name: String, body: Any) {
-        mockName = name
-        mockBody = body
-        super.init() // Calling the superclass initializer
-    }
-
-    override var name: String {
-        mockName
-    }
-
-    override var body: Any {
-        mockBody
-    }
-}
-
 final class IAFWebViewModelPreloadingTests: XCTestCase {
     // MARK: - setup
